Add microsoft store support
